Move schematic PIParameter into caliopen_pi.parameters
