+++ src/backend/main/py.main/caliopen_main/objects/pi.py
@@ -11,21 +11,9 @@ from uuid import UUID
 
 from cassandra.cqlengine import columns
 from elasticsearch_dsl import InnerObjectWrapper, Integer
-from schematics.models import Model
-from schematics.types import IntType, DateTimeType
 
 from caliopen_storage.store import BaseUserType
 from .base import ObjectIndexable
-
-
-class PIParameter(Model):
-    """The privacy indexes schematics parameter definition."""
-
-    technic = IntType()
-    comportment = IntType()
-    context = IntType()
-    version = IntType()
-    date_update = DateTimeType()
 
 
 class PIModel(BaseUserType):
