+++ src/backend/main/py.pi/setup.py
@@ -15,6 +15,7 @@ with open(init_file) as v_file:
         match(v_file.read()).group(1)
 
 requires = [
+    'schematics',
     'pgpy',
 ]
